requirements.txt
av==10.0.0
cvzone==1.6.1
numpy==1.26.2
opencv_contrib_python==4.8.1.78
opencv_python==4.8.0.76
Pillow==10.0.1
Pillow==10.1.0
pygame==2.5.2
streamlit==1.28.0
streamlit_webrtc==0.47.1
ultralytics==8.0.196